+++ src/components/TableView/TableView.tsx
@@ -63,15 +63,6 @@ export const TableView: React.FC<Props> = ({ data, id, options, width, height, e
   });
 
   /**
-   * Sticky position
-   */
-  const { containerRef, style } = useContentPosition({
-    width,
-    height,
-    sticky: options.sticky,
-  });
-
-  /**
    * Content Sizes
    */
   const {
@@ -82,6 +73,16 @@ export const TableView: React.FC<Props> = ({ data, id, options, width, height, e
     tableHeaderRef,
     tableContentTopOffset,
   } = useContentSizes({ height, options, tableData });
+
+  /**
+   * Sticky position
+   */
+  const { containerRef, style } = useContentPosition({
+    width,
+    height,
+    sticky: options.sticky,
+    scrollableContainerRef,
+  });
 
   /**
    * First selected row ref